==> notebooks/02_descriptive_plots_and_tables.nb.Rmd <==
---
title: "02-Tables and Plots R Notebook"
author: "AJ Strauman-Scott"
date: "`r Sys.Date()`"
output: html_notebook
---

This notebook performs Exploratory Data Analysis on the motor vehicle crashes data aggregated with the ACS data

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE)

# Load Libraries
library(tidyverse)
library(reshape2)
library(corrplot)
library(sf)
library(patchwork)
```

## LOAD DATA

```{r load-data}
# Load Data
df <- readRDS("../data/models/social-risk-crash-rate-data.rds")
glimpse(df)
```
```{r stats}
summary(df)
```

```{r summary_stats}
summary_vars <- c(
  "crash_rate_per_1000", "injury_rate_per_1000", "fatality_rate_per_1000",
  "median_income", "pct_below_poverty", "median_gross_rent",
  "pct_vehicle", "pct_public_transit", "pct_walk", "pct_bike",
  "poverty_vehicle_interaction", "unemployment_vehicle_interaction"
)

summary_table <- df %>%
  select(all_of(summary_vars)) %>%
  summarise(across(everything(), list(
    Mean = ~ mean(.x, na.rm = TRUE),
    SD   = ~ sd(.x, na.rm = TRUE),
    Max  = ~ max(.x, na.rm = TRUE)
  ), .names = "{.col}_{.fn}")) %>%
  pivot_longer(cols = everything(),
               names_to = c("Variable", ".value"),
               names_pattern = "(.*)_(Mean|SD|Max)") %>%
  arrange(Variable)

# Print a clean table
knitr::kable(summary_table, digits = 2,
             caption = "Summary Statistics of Key Variables (2018–2023)")

```

```{r time-series}
# Time Series Trend Plot

# Compute annual mean crash rate by borough
annual_crash_trend <- df %>%
      filter(borough != "Unknown") %>%  
  group_by(year, borough) %>%
  summarise(mean_crash_rate = mean(crash_rate_per_1000, na.rm = TRUE), .groups = "drop")

# Plot
timeseries <- ggplot(annual_crash_trend, aes(x = year, y = mean_crash_rate, color = borough, group = borough)) +
  geom_line(size = 1.2) +
  geom_point(size = 2) +
  theme_minimal() +
  labs(
    title = "Annual Mean Crash Rate per 1K Residents by Borough",
    x = "Year",
    y = "Mean Crash Rate per 1,000 Residents",
    color = "Borough"
  )

ggsave("../report/plots/crash_rate_by_borough_plot.png", timeseries, width = 10, height = 6, dpi = 300)
```
```{r corr-heatmap}
# Correlation Heatmap

corr_vars <- df %>%
  select(crash_rate_per_1000, median_income, pct_below_poverty, 
         median_gross_rent, pct_vehicle, pct_public_transit)

# Compute correlation matrix
corr_matrix <- cor(corr_vars, use = "pairwise.complete.obs")

# Plot and save the correlation heatmap
png("../report/plots/key_variables_corr_plot.png", width = 10, height = 6, units = "in", res = 300)

corrplot(corr_matrix,
         method = "color",
         type = "upper",
         tl.col = "black",
         tl.srt = 45,
         addCoef.col = "black",
         number.cex = 0.7,
         title = "Correlation Heatmap of Key Variables",
         mar = c(0, 0, 1, 0))

dev.off()
```

```{r geospatial}
# Load and Prep Data
tracts_sf <- st_read("../data/shapefiles/tl_2020_36_tract.shp") %>%
  filter(COUNTYFP %in% c("005", "047", "061", "081", "085"))

# Prepare data for 2018 and 2022
df_2018 <- df %>%
  filter(year == 2018, borough %in% c("Bronx", "Brooklyn", "Manhattan", "Queens", "Staten Island")) %>%
  select(geoid, crash_rate_per_1000)

df_2022 <- df %>%
  filter(year == 2022, borough %in% c("Bronx", "Brooklyn", "Manhattan", "Queens", "Staten Island")) %>%
  select(geoid, crash_rate_per_1000)


# Merge crash data with spatial data
map_2018 <- tracts_sf %>%
  left_join(df_2018, by = c("GEOID" = "geoid"))

map_2022 <- tracts_sf %>%
  left_join(df_2022, by = c("GEOID" = "geoid"))

# SEt global min and max for scale
crash_min <- 0
crash_max <- 10

# Plot Heatmaps
map2018 <- ggplot(map_2018) +
  geom_sf(aes(fill = crash_rate_per_1000), color = NA) +
  scale_fill_viridis_c(option = "plasma", na.value = "grey80", limits = c(crash_min, crash_max)) +
  theme_minimal() +
  labs(
    title = "Crash Rate per 1,000 Residents (2018)",
    fill = "Crashes / 1,000 Residents"
  )

map2022 <- ggplot(map_2022) +
  geom_sf(aes(fill = crash_rate_per_1000), color = NA) +
  scale_fill_viridis_c(option = "plasma", na.value = "grey80", limits = c(crash_min, crash_max)) +
  theme_minimal() +
  labs(
    title = "Crash Rate per 1,000 Residents (2022)",
    fill = "Crashes / 1,000 Residents"
  )
print(map2018)
print(map2022)

ggsave("../report/plots/2018_map_plot.png", map2018, width = 10, height = 6, dpi = 300)
ggsave("../report/plots/2022_map_plot.png", map2022, width = 10, height = 6, dpi = 300)

combined_maps <- map2018 + map2022 + plot_layout(guides = "collect") & theme(legend.position = "bottom")

# Save combined map
ggsave("../report/plots/2018_2022_comparison.png", combined_maps, width = 12, height = 6, dpi = 300)

```
